move is_fuchsian into differential_operator_1_1
